embed -> parameters
When inserting a video via an embed code, we should place relevant parameters from the iframe URL in an object called "parameters" rather than "embed". This follows the convention from our other video pickers, and also adds rooms for an "embed" object in which the user may overwrite the "parameters".